refactor(product page): Move to correct path and drop unused stuff
