+++ frontend/app/(defaultLayout)/products/[handle]/json-ld-scripts.tsx
