+++ frontend/app/(defaultLayout)/products/[handle]/page.tsx
