--- frontend/pages/products/[handle].tsx
@@ -1,2 +1,0 @@
-export { default } from '@templates/product';
-export { getServerSideProps } from '@templates/product/server';
