+++ frontend/templates/product/hooks/useSelectedVariant.ts
@@ -66,11 +66,3 @@ export function useSelectedVariant(
       return getProductVariantURI(searchParamVariantId);
    }
 }
-
-export function useDefaultVariantId(product: Product): string {
-   const variant =
-      product.variants.find(
-         (variant) => variant.quantityAvailable && variant.quantityAvailable > 0
-      ) ?? product.variants[0];
-   return variant.id;
-}
